Move XDC circuit paths to config
We would like to be able to tell what the pre- and post-link circuit
paths of the design are going to be without having to elaborate the
`PlatformShim` to get at the corresponding `XDCPathToCircuitAnnotation`.

This commit adds the two new config options `PreLinkCircuitPath` and
`PostLinkCircuitPath`, and moves the platform-specific paths out from
`F1Shim` and `VitisShim` and into the corresponding `F1Config` and
`VitisConfig`, respectively. This allows other parts of the codebase to
determine these paths purely from the config. Since the
platform-specific shim code now no longer has to specify these paths
directly, the `SpecifyXDCCircuitPaths` is moved up into the parent
`PlatformShim` class which annotates the paths provided by the config.